requirements.txt
anyio==4.7.0
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
python-telegram-bot==21.9
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
